api-cop16/requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sklearn==0.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
openai==0.28.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
